+++ apps/www/src/app/(main)/(docs)/components/editor/examples/invalid.tsx
@@ -1,4 +1,4 @@
-import { Editor } from "@jamsr-ui/react";
+import { Editor } from "@jamsr-ui/editor";
 
 export const EditorInvalid = () => {
   return <Editor isInvalid helperText="Description is required" />;