+++ apps/www/src/app/(main)/(docs)/components/editor/examples/usage.tsx
@@ -1,4 +1,4 @@
-import { Editor } from "@jamsr-ui/react";
+import { Editor } from "@jamsr-ui/editor";
 
 export const EditorUsage = () => {
   return <Editor label="Enter product description" />;