+++ packages/components/dialog/src/index.ts
@@ -6,4 +6,4 @@ export { DialogContent, type DialogContentProps } from "./dialog-content";
 export { DialogFooter, type DialogFooterProps } from "./dialog-footer";
 export { DialogHeader, type DialogHeaderProps } from "./dialog-header";
 export { DialogTrigger } from "./dialog-trigger";
-export { DialogState } from "./dialog-state";
+export { DialogState, useDialogState } from "./dialog-state";