+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/popover
 
+## 0.0.89
+
+### Patch Changes
+
+- fix dialog tooltip popover
+
 ## 0.0.88
 
 ### Patch Changes