+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/select
 
+## 0.0.112
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/popover@0.0.89
+
 ## 0.0.111
 
 ### Patch Changes